refactor(rust): set default_node_name using clap
